requirements.txt
# Core CrewAI CLI + framework
crewai

# OpenAI client for GPT-4o-mini calls
openai

# HTTP client (used in any custom helpers)
requests

# Crawl4AI for your scraper (AsyncWebCrawler, JsonCssExtractionStrategy, etc.)
crawl4ai

# YAML parsing (for reading configs)
PyYAML

# Data handling (CSV/JSON ingestion, DataFrame manipulation)
pandas

# PDF generation (ReportLab)
reportlab

# FastAPI + Uvicorn for the optional REST API export
fastapi
uvicorn

# (Optional) If you ever need to load .env files:
python-dotenv
